Make generators output float64 and convert to proper bit depth in the mixer for smoother gain control

+++ generators/generator.go
@@ -3,7 +3,7 @@ package generators
 import "github.com/bspaans/bs8bs/audio"
 
 type Generator interface {
-	GetSamples(cfg *audio.AudioConfig, n int) []int
+	GetSamples(cfg *audio.AudioConfig, n int) []float64 // return samples between -1.0 and 1.0
 	SetPitch(float64)
 }
